simple-policy-assistant/requirements.txt
google-generativeai==0.3.2
PyPDF2==3.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
numpy==1.24.3